gcp/esame-14-giugno-2023/src/requirements.txt
flask
flask-restful
WTForms
requests
google-cloud-firestore
google-cloud-pubsub
gunicorn
python-dateutil
